Add clang-format and cppcheck for firmware, fix printf format warnings
Set up clang-format (.clang-format) matching existing Stroustrup style
(4-space indent, function braces on next line, aligned declarations).
Added cppcheck with suppression file for HAL/lwIP false positives.
New Makefile targets: format, format-check, lint.

Fixed 6 printf format warnings: %lX requires unsigned long but uint32_t
is unsigned int on ARM — added (unsigned long) casts in PHY register
debug output (eth.c, shell.cpp).

+++ firmware/src/net/eth.c
@@ -259,7 +259,7 @@ void eth_poll_link(void)
     bool phy_link = (regval & PHY_BSR_LINK_STATUS) != 0;
 
     if (phy_link && !prev_link) {
-        printf("[eth] PHY link UP (BSR=0x%04lX)\r\n", regval);
+        printf("[eth] PHY link UP (BSR=0x%04lX)\r\n", (unsigned long)regval);
 
         /* Read negotiated speed/duplex from PHY SCSR */
         uint32_t scsr = 0;
